เพิ่มผู้สร้างสามารถลบหรืออัพเดทเฉพาะข้อมูลของตัวเอง และเห็น edit delete

+++ backend/routes/posts.js
@@ -32,7 +32,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then((result) => {
     console.log(result);
@@ -59,12 +60,17 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single("image"), (req
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: ImagePath
+    imagePath: ImagePath,
+    creator: req.userData.userId
   });
   console.log("update result " + post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Update success' });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update success' });
+    } else {
+      res.status(401).json({ message: "Not authorizad!" });
+    }
+
   });
 });
 
@@ -105,9 +111,12 @@ router.get("/:id", (req , res, next) => {
 
 router.delete('/:id', checkAuth, (req, res, next) => {
   console.warn('ID has been deleted ' + req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Post deleted!' });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Delete success" });
+    } else {
+      res.status(401).json({ message: "Not authorizad!" });
+    }
   });
 });
